Fixes meshes not updating until save is reloaded.
